% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certepackage.R
\docType{package}
\name{certedb-package}
\alias{certedb}
\alias{certedb-package}
\title{certedb: A Certe R Package for Connecting to Databases}
\description{
A Certe R Package for connecting to internal Certe databases. This package is part of the 'certedata' universe.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://certe-medical-epidemiology.github.io/certedb}
  \item \url{https://github.com/certe-medical-epidemiology/certedb}
}

}
\author{
\strong{Maintainer}: Matthijs S. Berends \email{m.berends@certe.nl}

Authors:
\itemize{
  \item Erwin E. A. Hassing \email{e.hassing@certe.nl}
}

Other contributors:
\itemize{
  \item Certe Medical Diagnostics & Advice Foundation [copyright holder, funder]
}

}
\keyword{internal}
