% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certedb.R
\name{certedb}
\alias{certedb}
\alias{certedb_close}
\title{Connect to Certe Database}
\usage{
certedb(
  dbname = "certemmb",
  host = NULL,
  port = NULL,
  username = NULL,
  password = NULL,
  driver = get_driver(dbname = dbname),
  info = TRUE
)

certedb_close(conn, ...)
}
\arguments{
\item{dbname}{database name}

\item{host, port, username, password}{credentials}

\item{driver}{database driver to use}

\item{info}{Default is \code{FALSE}. Display of connecting info.}

\item{conn}{connection to close, such as the output of \code{\link[=certedb]{certedb()}}}

\item{...}{arguments passed on to \code{\link[DBI:dbDisconnect]{DBI::dbDisconnect()}}}
}
\description{
Connect to a remote database using \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.
}
