% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mmb.R
\name{certedb_check_tbls}
\alias{certedb_check_tbls}
\title{Tabellen valideren in MySQL-/MariaDB-database}
\usage{
certedb_check_tbls(tbls, con = NULL, dbname = "certemmb", info = TRUE)
}
\arguments{
\item{tbls}{Tabellen in de database die gecontroleerd moeten worden.}

\item{con}{Standaard is leeg, waarmee de verbinding gemaakt wordt op basis van de omgevingsvariabelen van de huidige gebruiker: \code{"DB_HOST"}, \code{"DB_PORT"}, \code{"DB_USERNAME"} en \code{"DB_PASSWORD"}.}

\item{dbname}{Standaard is \code{"certemmb"}. Naam van de database die geselecteerd moet worden. Wordt genegeerd als \code{con} al een bestaande verbinding is.}

\item{info}{Standaard is \code{TRUE}. Waarschuwing weergeven over niet-bestaande tabellen die vervangen worden door bestaande tabellen.}
}
\description{
Hiermee worden tabellen opgezocht in een MySQL-/MariaDB-database. Niet-bestaande tabellen worden hoofdletterongevoelig vergeleken met bestaande tabellen. Wanneer een niet-bestaande tabel wel voorkomt met een prefix eindigend op een underscore (\code{"_"}, wordt de tabelnaam overschreven door de bestaande tabelnaam die alfabetisch als laatste voorkomt.
}
