% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mmb.R
\name{certedb_checkmmb_mtrlcodes}
\alias{certedb_checkmmb_mtrlcodes}
\alias{certedb_checkmmb_testcodes}
\title{Controleren van nieuwe materiaal- en testcodes}
\usage{
certedb_checkmmb_mtrlcodes(...)

certedb_checkmmb_testcodes(info = TRUE, ...)
}
\arguments{
\item{...}{Parameters die doorgegeven worden aan \code{\link{certedb_query}}.}
}
\description{
Controleert originele materiaal- en testcodes en vergelijkt ze met de nieuwe \code{temporary_*}-tabellen.
}
