% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mmb.R
\name{certedb_close}
\alias{certedb_close}
\title{Disconnect (close) a connection}
\usage{
certedb_close(conn, ...)
}
\arguments{
\item{conn}{A \linkS4class{DBIConnection} object, as returned by
\code{\link[DBI:dbConnect]{dbConnect()}}.}

\item{...}{Other parameters passed on to methods.}
}
\value{
\code{dbDisconnect()} returns \code{TRUE}, invisibly.
}
\description{
This closes the connection, discards all pending work, and frees
resources (e.g., memory, sockets).

\Sexpr[results=rd,stage=render]{DBI:::methods_as_rd("dbDisconnect")}
}
\section{Specification}{



A warning is issued on garbage collection when a connection has been
released without calling \code{dbDisconnect()},
but this cannot be tested automatically.
A warning is issued immediately when calling \code{dbDisconnect()} on an
already disconnected
or invalid connection.


}

\examples{
con <- dbConnect(RSQLite::SQLite(), ":memory:")
dbDisconnect(con)
}
\seealso{
\code{\link{certedb_query}} voor het direct gebruik van query's.
}
