% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mmb.R
\name{certedb_query}
\alias{certedb_query}
\title{SQL-query uitvoeren op MySQL-/MariaDB-database}
\usage{
certedb_query(
  query,
  limit = 1e+07,
  con = NULL,
  dbname = "certemmb",
  info = TRUE,
  check_mtrl_test_codes = TRUE,
  binary_as_logical = TRUE,
  auto_append_prefix = TRUE,
  auto_transform = TRUE,
  datenames = "en",
  dateformat = "\%Y-\%m-\%d",
  timeformat = "\%H:\%M",
  decimal.mark = ".",
  big.mark = "",
  timezone = "UTC",
  na = c("", "NULL", "NA"),
  ...
)
}
\arguments{
\item{query}{(Bestand met) SQL-tekst die uitgevoerd moet worden. Tabelnamen in de query hoeven geen \code{"temporary_"} of \code{"certemm_"} te bevatten en zijn hoofdletterongevoelig. Deze query wordt bij het object opgeslagen als eigenschap \code{query} dat bekeken kan worden met de functie \code{\link{qry}}.}

\item{limit}{Standaard is 10.000.000. Het aantal rijen dat maximaal opgehaald moet worden.}

\item{con}{Standaard is leeg, waarmee de verbinding gemaakt wordt op basis van de omgevingsvariabelen van de huidige gebruiker: \code{"DB_HOST"}, \code{"DB_PORT"}, \code{"DB_USERNAME"} en \code{"DB_PASSWORD"}.}

\item{dbname}{Standaard is \code{"certemmb"}. Naam van de database die geselecteerd moet worden. Wordt genegeerd als \code{con} al een bestaande verbinding is.}

\item{info}{Standaard is \code{TRUE}. Printen van voortgang van run/fetch en het uiteindelijke aantal rijen en kolommen dat gedownload is.}

\item{check_mtrl_test_codes}{Standaard is \code{TRUE}. Controleren van materiaal- en testcodes in de tijdelijke tabellen van de database. Hiervoor moet ook \code{info = TRUE} zijn.}

\item{binary_as_logical}{Standaard is \code{TRUE}. Kolommen die alleen de waarden \code{0} en/of \code{1} bevatten, tranformeren naar logical m.b.v. \code{\link{tbl_binary2logical}}.}

\item{auto_append_prefix}{Standaard is \code{TRUE}. Wanneer tabellen die voorkomen in de query niet bestaan, wordt gezocht naar tabellen met dezelfde naam die beginnen met \code{"temporary_"} of \code{"certemm_"}. Wanneer zo'n tabel gevonden, wordt die gebruikt.}

\item{auto_transform}{Standaard is \code{TRUE}. Automatisch alle gedownloade kolommen transformeren met \code{\link{tbl_guess_columns}}.}

\item{...}{Overige ongebruikte parameters}
}
\description{
Een SQL-query uitvoeren op een MySQL-/MariaDB-database van bijv. Certe. De output krijgt een \code{qry}-attribuut, dat met \code{\link{qry}} opgehaald kan worden.
}
\examples{
\dontrun{
locaties <- certedb_query("SELECT * FROM temporary_certemm_locaties")
locaties <- certedb_query("SELECT * FROM temporary_certemm_locaties", limit = 5)

# Door `auto_append_prefix = TRUE` wordt:
locaties <- certedb_query("SELECT * FROM locaties")
# vertaald naar:
locaties <- certedb_query("SELECT * FROM temporary_certemm_locaties")
}
}
\seealso{
\code{\link{certedb}} voor alleen het verbinden met de Certe-databaseserver en \code{\link{certedb_getmmb}} om ineens alle relevante MMB-gegevens te downloaden.
}
