% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mmb.R
\name{certedb_tbls}
\alias{certedb_tbls}
\title{Tabellen ophalen uit MySQL-/MariaDB-database}
\usage{
certedb_tbls(con = NULL, dbname = "certemmb")
}
\arguments{
\item{con}{Standaard is leeg, waarmee de verbinding gemaakt wordt op basis van de omgevingsvariabelen van de huidige gebruiker: \code{"DB_HOST"}, \code{"DB_PORT"}, \code{"DB_USERNAME"} en \code{"DB_PASSWORD"}.}

\item{dbname}{Standaard is \code{"certemmb"}. Naam van de database die geselecteerd moet worden. Wordt genegeerd als \code{con} al een bestaande verbinding is.}
}
\value{
\code{\link{list}} met alle gevonden SQL-tabellen.
}
\description{
Alle tabellen uit een MySQL-/MariaDB-database zoals de Certe-database ophalen en in een \code{\link{list}} plaatsen, waarbij elke waarde in de lijst ook de naam is van het element in de lijst.
}
\seealso{
\code{\link{certedb_query}} voor het direct gebruik van query's.
}
