% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/di_env.R
\docType{data}
\name{gl}
\alias{gl}
\alias{di}
\title{List of DI Columns}
\format{
An object of class \code{list} of length 78.

An object of class \code{list} of length 151.
}
\usage{
gl

di
}
\description{
List of DI Columns
}
\keyword{datasets}
\keyword{internal}
