% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certedb.R
\docType{data}
\name{db_connect}
\alias{db_connect}
\alias{db_close}
\alias{db}
\title{Connect to Certe Database}
\usage{
db_connect(
  dbname = "certemmb",
  host = NULL,
  port = NULL,
  username = NULL,
  password = NULL,
  driver = get_driver(dbname = dbname),
  info = TRUE
)

db_close(conn, ...)

db
}
\arguments{
\item{dbname}{database name}

\item{host, port, username, password}{credentials}

\item{driver}{database driver to use}

\item{info}{Default is \code{FALSE}. Display of connecting info.}

\item{conn}{connection to close, such as the output of \code{\link[=db_connect]{db_connect()}}}

\item{...}{arguments passed on to \code{\link[DBI:dbDisconnect]{DBI::dbDisconnect()}}}
}
\description{
Connect to a remote database using \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.
}
\details{
\code{db} is a \link{list} with all the database tables and columns.
}
\keyword{datasets}
