% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/di.R
\docType{data}
\name{di}
\alias{di}
\title{List with Variables}
\format{
A \link{list} with 122 items.
}
\usage{
di
}
\description{
A list with variables to be used with \code{\link[=get_diver_data]{get_diver_data()}}.
}
\keyword{datasets}
