% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preset.R
\name{preset}
\alias{preset}
\alias{preset.list}
\alias{preset.thisfolder.list}
\alias{preset.exists}
\alias{preset.read}
\alias{preset.add}
\alias{preset.thisfolder}
\title{Voorgedefinieerde variabelen selecteren}
\usage{
preset.list(fullname = FALSE, recursive = FALSE)

preset.thisfolder.list(fullname = FALSE, recursive = TRUE)

preset.exists(name, returnlocation = FALSE)

preset.read(name)

preset.add(name, vector)

preset.thisfolder(recursive = TRUE)
}
\arguments{
\item{fullname}{Standaard is \code{FALSE}. Bestandsnamen weergeven met prefix \code{prefix_} en suffix \code{.sql}.}

\item{recursive}{Ook onderliggende mappen doorzoeken.}

\item{name}{Tekst(vector) om te zoeken in de map \code{Sys.getenv("R_REFMAP")}.}

\item{vector}{Tekst(vector) met kolomnamen die opgeslagen moet worden als nieuwe preset. Deze kunnen geselecteerd worden met \code{\link{db}}.}
}
\description{
Dit downloadt bestanden uit de map \code{Sys.getenv("R_REFMAP")} die de structuur \code{preset_*.sql} hebben. Gebruik een of meerdere van deze (zelf te maken) bestanden als parameter \code{select_preset} in de functies \code{\link{certedb_getmmb}} en \code{\link{certedb_getmmb_tat}}. Zie Details.
}
\details{
Gebruik: \cr
\itemize{
\item \code{preset.list} voor het weergeven van alle presets in de map \code{Sys.getenv("R_REFMAP")}; \cr
\item \code{preset.add} voor het toevoegen van een nieuwe preset; \cr
\item \code{preset.read} voor het uitlezen van een preset; \cr
\item \code{preset.thisfolder.list} voor het weergeven van alle presets in de huidige map(pen); \cr
\item \code{preset.thisfolder} voor het weergeven van de eerst gevonden preset in de huidge map; \cr
\item \code{preset.exists} voor toetsen of een preset bestaat.
}
}
