% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presets.R
\name{presets}
\alias{presets}
\title{Available Presets for Diver Server}
\usage{
presets()
}
\description{
This returns a data.frame with available presets, as defined in the secrets YAML file under \code{db.presets}.
}
