% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry.R
\encoding{UTF-8}
\name{qry}
\alias{qry}
\alias{qry<-}
\alias{has_qry}
\alias{qry_beautify}
\alias{qry_isvalid}
\title{Query van een object}
\usage{
qry(x)

qry(x) <- value

has_qry(tbl)

qry_beautify(query)

qry_isvalid(query)
}
\arguments{
\item{x}{Object.}

\item{value}{Waarde om toe te kennen.}

\item{query}{Bestaande query als tekst.}
}
\description{
Hiermee kan snel de eigenschap (\emph{attribute}) \code{qry} van een object weergegeven worden of aan een object toegewezen worden. Deze eigenschap wordt altijd toegewezen aan de output van \code{\link{certedb_query}} en vóór toewijzing gevalideerd.
}
\details{
De functie \code{qry_beautify} wordt intern gebruikt om een query netjes op te maken door statements (zoals SELECT en FROM) als hoofdletters weer te geven, commentaren te verwijderen en enters toe te voegen. \cr
De functie \code{qry_isvalid} toetst of een query begint met een van de statements op \url{https://mariadb.com/kb/en/library/basic-sql-statements/}.
}
\examples{
\dontrun{
data <- certedb_query("SELECT * FROM certemm_ord LIMIT 1")
qry(data)

qry(data) <- NULL
qry(data)
}
}
