% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{db_list_tables}
\alias{db_drop_table}
\alias{db_write_table}
\alias{db_has_table}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:backend_dbplyr]{db_drop_table}}, \code{\link[dplyr:backend_dbplyr]{db_has_table}}, \code{\link[dplyr:backend_dbplyr]{db_list_tables}}, \code{\link[dplyr:backend_dbplyr]{db_write_table}}}
}}

