% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shiny_explore}
\alias{shiny_explore}
\title{Search cBase Interactively}
\usage{
shiny_explore(
  preset = read_secret("db.preset_default_shiny"),
  diver_cbase = NULL,
  diver_project = read_secret("db.diver_project"),
  diver_dsn = if (diver_testserver == FALSE) read_secret("db.diver_dsn") else
    read_secret("db.diver_dsn_test"),
  diver_testserver = FALSE,
  diver_tablename = "data"
)
}
\arguments{
\item{preset}{a preset to choose from \code{\link[=presets]{presets()}}. Will be ignored if \code{diver_cbase} is set, even if it is set to \code{NULL}. Be sure to read \link[=presets]{the documentation} on how to use presets, and to see in which order the YAML keys will be run.}

\item{diver_cbase, diver_project, diver_dsn, diver_testserver}{properties to set in \code{\link[=db_connect]{db_connect()}}. The \code{diver_cbase} argument will be based on \code{preset}, but can also be set to blank \code{NULL} to manually select a cBase in a popup window.}

\item{diver_tablename}{name of the database table to download data from. This is hard-coded by DI and should normally never be changed.}
}
\description{
Use this Shiny app to search a cBase. There is also an RStudio add-in.
}
