% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mmb.R
\name{abc}
\alias{abc}
\alias{def}
\title{Title of function}
\usage{
abc()

def(d)
}
\arguments{
\item{d}{text}
}
\description{
Short description
}
\examples{
a <- "example1"
b <- "example2"
paste(a, b)

head(mtcars)

1 + 1
}
