% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mmb.R
\name{certedb}
\alias{certedb}
\title{Verbinding maken met MySQL-/MariaDB-database}
\usage{
certedb(
  host = Sys.getenv("DB_HOST"),
  port = Sys.getenv("DB_PORT"),
  username = Sys.getenv("DB_USERNAME"),
  password = Sys.getenv("DB_PASSWORD"),
  dbname = "certemmb",
  info = FALSE
)
}
\arguments{
\item{host, port, username, password}{Standaard zijn omgevingsvariabelen van de huidige gebruiker: \code{"DB_HOST"} (bij ontbreken wordt dit \code{localhost}), \code{"DB_PORT"} (bij ontbreken wordt dit \code{3306}), \code{"DB_USERNAME"} en \code{"DB_PASSWORD"}. Inloggegevens voor de MySQL/MariaDB-server.}

\item{dbname}{Standaard is \code{"certemmb"}. Naam van de database die geselecteerd moet worden.}

\item{info}{Standaard is \code{FALSE}. Weergeven van informatie over het verbinden.}
}
\description{
Dit maakt verbinding met een MySQL-/MariaDB-database zoals de Certe-databaseserver.
}
\examples{
\dontrun{

# gegevens direct downloaden naar tibble:
locaties <- certe_db() \%>\%
  tbl("temporary_certemm_locaties") \%>\%
  collect()

# alleen query weergeven en niet uitvoeren:
locaties <- certe_db() \%>\%
  tbl("temporary_certemm_locaties") \%>\%
  show_query()

# rest werkt met dpylr, zoals filter() en select():
locaties <- certe_db() \%>\%
  tbl("temporary_certemm_locaties") \%>\%
  filter(zkhgroepcode > 0, cu_sd == "Noord") \%>\%
  select(instelling) \%>\%
  collect()

# gebruik head() zoals LIMIT in MySQL:
locaties <- certe_db() \%>\%
  tbl("temporary_certemm_locaties") \%>\%
  head(5) \%>\%
  collect()
}
}
\seealso{
\code{\link{certedb_query}} voor het direct gebruik van query's en \code{\link{certedb_close}} voor het sluiten van een verbinding.
}
