% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{certedb}
\alias{certedb}
\title{Connecting to the MySQL-/MariaDB-database}
\usage{
certedb(
  host = read_secret("db.mmb.host"),
  port = read_secret("db.mmb.port"),
  username = read_secret("db.mmb.username"),
  password = read_secret("db.mmb.password"),
  dbname = "certemmb",
  info = FALSE
)
}
\arguments{
\item{host, port, username, password}{Default is the current user's environment variables: \code{"db.mmb.host"} (if missing will be \code{localhost}), \code{"db.mmb.port"} (if missing will be \code{3306}), \code{"db.mmb.username"} and \code{"db.mmb.password"}. Login credentials for the MySQL/MariaDB-server.}

\item{dbname}{Default is \code{"certemmb"}. Name of the database to be selected.}

\item{info}{Default is \code{FALSE}. Display of connecting info.}
}
\description{
Makes a connection to a MySQL-/MariaDB-database like the Certe-databaseserver.
}
\examples{
\dontrun{

# download data directly to tibble:
locaties <- certe_db() \%>\%
  tbl("temporary_certemm_locaties") \%>\%
  collect()

# only show query without running it:
locaties <- certe_db() \%>\%
  tbl("temporary_certemm_locaties") \%>\%
  show_query()

# works with dpylr functions like filter() and select():
locaties <- certe_db() \%>\%
  tbl("temporary_certemm_locaties") \%>\%
  filter(zkhgroepcode > 0, cu_sd == "Noord") \%>\%
  select(instelling) \%>\%
  collect()

# use head() like LIMIT in MySQL:
locaties <- certe_db() \%>\%
  tbl("temporary_certemm_locaties") \%>\%
  head(5) \%>\%
  collect()
}
}
\seealso{
\code{\link{certedb_query}} for the use of query's and \code{\link{certedb_close}} for closing a connection.
}
