% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{certedb_check_tbls}
\alias{certedb_check_tbls}
\title{Verify tables in MySQL-/MariaDB-database}
\usage{
certedb_check_tbls(tbls, con = NULL, dbname = "certemmb", info = TRUE)
}
\arguments{
\item{tbls}{Tables in the database to be verified.}

\item{con}{Default is empty, by which the connection is made based on the current user's environment variables: \code{"db.mmb.host"}, \code{"db.mmb.port"}, \code{"db.mmb.username"} en \code{"db.mmb.password"}.}

\item{dbname}{Default is \code{"certemmb"}. Name of the database to be selected. Will be ignored if \code{con} is already an existing connection.}

\item{info}{Default is \code{TRUE}. Display a warning about non-existing tables to be replaced by existing tables.}
}
\description{
Checks for tables in a MySQL-/MariaDB-database. Non-existing tables will be compared (case insensitive) to existing tables. When a non-existing table occurs with a prefix ending with an underscore (\code{"_"}), the table name will be overwritten with an existing table name listed last in alphabetical order.
}
