% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{certedb_checkmmb_mtrlcodes}
\alias{certedb_checkmmb_mtrlcodes}
\alias{certedb_checkmmb_testcodes}
\title{Check for new material- and testcodes}
\usage{
certedb_checkmmb_mtrlcodes(...)

certedb_checkmmb_testcodes(info = TRUE, ...)
}
\arguments{
\item{...}{Arguments to be used in \code{\link{certedb_query}}.}
}
\description{
Check original material- and testcodes and compares them to the new \code{temporary_*}-tables.
}
