% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{certedb_query}
\alias{certedb_query}
\title{Run SQL-query on a MySQL-/MariaDB-database}
\usage{
certedb_query(
  query,
  limit = 1e+07,
  con = NULL,
  dbname = "certemmb",
  info = TRUE,
  check_mtrl_test_codes = TRUE,
  binary_as_logical = TRUE,
  auto_append_prefix = TRUE,
  auto_transform = TRUE,
  datenames = "en",
  dateformat = "\%Y-\%m-\%d",
  timeformat = "\%H:\%M",
  decimal.mark = ".",
  big.mark = "",
  timezone = "UTC",
  na = c("", "NULL", "NA"),
  ...
)
}
\arguments{
\item{query}{(File containing) SQL-text to be run. Table names in the query don't need to contain \code{"temporary_"} or \code{"certemm_"} and are not case sensitive. This query will be saved with the object as attribute \code{query} which can be viewed with the \code{\link{qry}} function.}

\item{limit}{Default is 10.000.000. The maximum amount of rows to be retrieved.}

\item{con}{Default is empty, by which the connection is made based on the current user's environment variables: \code{"db.mmb.host"}, \code{"db.mmb.port"}, \code{"db.mmb.username"} en \code{"db.mmb.password"}.}

\item{dbname}{Default is \code{"certemmb"}. Name of the database to be selected. Will be ignored if \code{con} is already an existing connection.}

\item{info}{Default is \code{TRUE}. Printing of run/fetch progress en the resulting number of rows and columns downloaded.}

\item{check_mtrl_test_codes}{Default is \code{TRUE}. Check on material- and testcodes in the database's temporary tables. Also requires \code{info = TRUE}.}

\item{binary_as_logical}{Default is \code{TRUE}. Transform columns only containing \code{0} and/or \code{1} values to logical using \code{\link{tbl_binary2logical}}.}

\item{auto_append_prefix}{Default is \code{TRUE}. When the selected columns in query are non-existent, eponymous tables starting with \code{"temporary_"} or \code{"certemm_"}. When such a table is found, it will be used instead.}

\item{auto_transform}{Default is \code{TRUE}. Automatically transform all downloaded columns with \code{\link{tbl_guess_columns}}.}

\item{...}{not used}
}
\description{
Run an SQL-query on a MySQL-/MariaDB-database from, for example, Certe. Output gets a \code{qry}-attribute, which can be retrieved with \code{\link{qry}}.
}
\examples{
\dontrun{
locaties <- certedb_query("SELECT * FROM temporary_certemm_locaties")
locaties <- certedb_query("SELECT * FROM temporary_certemm_locaties", limit = 5)

# With `auto_append_prefix = TRUE`:
locaties <- certedb_query("SELECT * FROM locaties")
# will be transformed to:
locaties <- certedb_query("SELECT * FROM temporary_certemm_locaties")
}
}
\seealso{
\code{\link{certedb}} for connecting to the Certe-databaseserver and \code{\link{certedb_getmmb}} for downloading all relevant MMB data.
}
