% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{certedb_tbls}
\alias{certedb_tbls}
\title{Retrieve tables from MySQL-/MariaDB-database}
\usage{
certedb_tbls(con = NULL, dbname = "certemmb")
}
\arguments{
\item{con}{Default is empty, by which the connection is made based on the current user's environment variables: \code{"db.mmb.host"}, \code{"db.mmb.port"}, \code{"db.mmb.username"} en \code{"db.mmb.password"}.}

\item{dbname}{Default is \code{"certemmb"}. Name of the database to be selected. Will be ignored if \code{con} is already an existing connection.}
}
\value{
\code{\link{list}} containing all retrieved SQL-tables.
}
\description{
Retrieve all tables from a MySQL-/MariaDB-database like the Certe-database ophalen and place them in a \code{\link{list}}, where each listed value is also the name of the element in the list.
}
\seealso{
\code{\link{certedb_query}} for using query's.
}
