% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\docType{data}
\name{db_connect}
\alias{db_connect}
\alias{db_close}
\alias{db2}
\title{Connect to a Certe Database}
\usage{
db_connect(driver, ..., info = TRUE)

db_close(conn, ...)

db2
}
\arguments{
\item{driver}{database driver to use}

\item{...}{arguments passed on to \code{\link[DBI:dbDisconnect]{DBI::dbDisconnect()}}}

\item{info}{a logical to indicate whether info about the connection should be printed}

\item{conn}{connection to close, such as the output of \code{\link[=db_connect]{db_connect()}}}
}
\description{
Connect to an internal (yet remote) Certe database server using \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.
}
\details{
\code{db2} is a \link{list} with all the database tables and columns.
}
\keyword{datasets}
