% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diver_data.R
\name{get_diver_data}
\alias{get_diver_data}
\title{Download Data from Diver Server}
\usage{
get_diver_data(
  date_range = c(paste0(format(Sys.Date(), "\%Y"), "-01-01"), paste0(format(Sys.Date(),
    "\%Y"), "-12-31")),
  where = NULL,
  diver_project = read_secret("db.diver_project"),
  diver_cbase = read_secret("db.diver_cbase"),
  review_qry = interactive()
)
}
\arguments{
\item{date_range}{date range, can be length 1 or 2 to filter on the column \code{Ontvangstdatum}), or \code{NULL} to not filter on any date column. Defaults to current year.}

\item{where}{arguments to filter data on, will be passed on to \code{\link[dplyr:filter]{filter()}}. \strong{Do not use \code{&&} or \code{||} but only \code{&} or \code{|} in filtering.}}

\item{diver_project}{name of the Diver project}

\item{diver_cbase}{name of the Diver cbase, can be left blank to select using a popup window}

\item{review_qry}{a \link{logical} to indicate whether the query must be reviewed first, defaults to \code{TRUE} in interactive mode and \code{FALSE} otherwise}
}
\description{
Download Data from Diver Server
}
