% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diver_data.R, R/get_duckdb_data.R,
%   R/get_molis_data.R
\name{get_diver_data}
\alias{get_diver_data}
\alias{certedb_query}
\alias{get_duckdb_data}
\alias{certedb_getmmb}
\alias{certedb_getmmb_tat}
\title{Download Data from a Local or Remote Database}
\usage{
get_diver_data(
  date_range = this_year(),
  where = NULL,
  review_qry = interactive(),
  antibiogram_type = "sir",
  distinct = TRUE,
  auto_transform = TRUE,
  preset = read_secret("db.preset_default"),
  diver_cbase = NULL,
  diver_project = read_secret("db.diver_project"),
  diver_dsn = if (diver_testserver == FALSE) read_secret("db.diver_dsn") else
    read_secret("db.diver_dsn_test"),
  diver_testserver = FALSE,
  info = interactive()
)

certedb_query(query, where = NULL, auto_transform = TRUE, info = interactive())

get_duckdb_data(
  date_range = this_year(),
  where = NULL,
  preset = NULL,
  review_qry = interactive(),
  duckdb_path = read_secret("db.duckdb_path"),
  duckdb_table = read_secret("db.duckdb_table"),
  auto_transform = TRUE,
  info = interactive()
)

certedb_getmmb(
  dates = NULL,
  where = NULL,
  select_preset = "mmb",
  preset = "mmb",
  select = NULL,
  add_cols = NULL,
  info = interactive(),
  first_isolates = FALSE,
  eucast_rules = "all",
  keep_synonyms = getOption("AMR_keep_synonyms", FALSE),
  mic = FALSE,
  disk = FALSE,
  zipcodes = FALSE,
  ziplength = 4,
  tat_hours = FALSE,
  only_real_patients = TRUE,
  only_conducted_tests = TRUE,
  only_validated = FALSE,
  only_show_query = FALSE,
  review_where = interactive(),
  auto_transform = TRUE,
  query = NULL,
  ...
)

certedb_getmmb_tat(
  dates = NULL,
  where = NULL,
  add_cols = NULL,
  info = interactive(),
  only_real_patients = TRUE,
  only_conducted_tests = TRUE,
  only_validated = TRUE,
  only_show_query = FALSE,
  ...
)
}
\arguments{
\item{date_range}{date range, can be length 1 or 2 (or more to use the min/max) to filter on the column \code{Ontvangstdatum}. Defaults to \code{\link[=this_year]{this_year()}}. Use \code{NULL} to set no date filter. Can also be years, or functions such as \code{\link[certetoolbox:days_around_today]{last_month()}}.}

\item{where}{arguments to filter data on, will be passed on to \code{\link[dplyr:filter]{filter()}}. \strong{Do not use \code{&&} or \code{||} but only \code{&} or \code{|} in filtering.}}

\item{review_qry}{a \link{logical} to indicate whether the query must be reviewed first, defaults to \code{TRUE} in interactive mode and \code{FALSE} otherwise. This will always be \code{FALSE} in Quarto / R Markdown, since the output of \code{\link[knitr:output_type]{knitr::pandoc_to()}} must be \code{NULL}.}

\item{antibiogram_type}{antibiotic transformation mode. Leave blank to strip antibiotic results from the data, \code{"sir"} to keep SIR values, \code{"mic"} to keep MIC values or \code{"disk"} to keep disk diffusion values. Values will be cleaned with \code{\link[AMR:as.sir]{as.sir()}}, \code{\link[AMR:as.mic]{as.mic()}} or \code{\link[AMR:as.disk]{as.disk()}}.}

\item{distinct}{\link{logical} to apply \code{\link[=distinct]{distinct()}} to the resulting data set}

\item{auto_transform}{\link{logical} to apply \code{\link[=auto_transform]{auto_transform()}} to the resulting data set}

\item{preset}{a preset to choose from \code{\link[=presets]{presets()}}. Will be ignored if \code{diver_cbase} is set, even if it is set to \code{NULL}.}

\item{diver_cbase, diver_project, diver_dsn, diver_testserver}{properties to set in \code{\link[=db_connect]{db_connect()}}. The \code{diver_cbase} argument will be based on \code{preset}, but can also be set to blank \code{NULL} to manually select a cBase in a popup window.}

\item{info}{settings for old \code{certedb_getmmb()} function}

\item{query}{a \link{data.frame} to view the query of, or a \link{character} string to run as query in \code{\link[=certedb_getmmb]{certedb_getmmb()}} (which will ignore all other arguments, except for \code{where}, \code{auto_transform} and \code{info}).}

\item{duckdb_path}{path to the local DuckDB database}

\item{duckdb_table}{name of the DuckDB database to retrieve data from}

\item{dates}{date range, can be length 1 or 2 (or more to use the min/max) to filter on the column \code{Ontvangstdatum}. Defaults to \code{\link[=this_year]{this_year()}}. Use \code{NULL} to set no date filter. Can also be years, or functions such as \code{\link[certetoolbox:days_around_today]{last_month()}}.}

\item{select_preset}{settings for old \code{certedb_getmmb()} function}

\item{select}{settings for old \code{certedb_getmmb()} function}

\item{add_cols}{settings for old \code{certedb_getmmb()} function}

\item{first_isolates}{settings for old \code{certedb_getmmb()} function}

\item{eucast_rules}{settings for old \code{certedb_getmmb()} function}

\item{keep_synonyms}{a \link{logical} to indicate if old, previously valid taxonomic names must be preserved and not be corrected to currently accepted names. The default is \code{FALSE}, which will return a note if old taxonomic names were processed. The default can be set with the \link[AMR:AMR-options]{package option} \code{\link[AMR:AMR-options]{AMR_keep_synonyms}}, i.e. \code{options(AMR_keep_synonyms = TRUE)} or \code{options(AMR_keep_synonyms = FALSE)}.}

\item{mic}{settings for old \code{certedb_getmmb()} function}

\item{disk}{settings for old \code{certedb_getmmb()} function}

\item{zipcodes}{settings for old \code{certedb_getmmb()} function}

\item{ziplength}{settings for old \code{certedb_getmmb()} function}

\item{tat_hours}{settings for old \code{certedb_getmmb()} function}

\item{only_real_patients}{settings for old \code{certedb_getmmb()} function}

\item{only_conducted_tests}{settings for old \code{certedb_getmmb()} function}

\item{only_validated}{settings for old \code{certedb_getmmb()} function}

\item{only_show_query}{settings for old \code{certedb_getmmb()} function}

\item{review_where}{a \link{logical} to indicate whether the query must be reviewed first, defaults to \code{TRUE} in interactive mode and \code{FALSE} otherwise. This will always be \code{FALSE} in Quarto / R Markdown, since the output of \code{\link[knitr:output_type]{knitr::pandoc_to()}} must be \code{NULL}.}

\item{...}{not used anymore, previously settings for old \code{certetools::certedb_getmmb()} function}
}
\description{
Thes functions can be used to download local or remote database data, e.g. Spectre data from DiveLine on a Diver server (from \href{https://www.dimins.com}{Dimensional Insight}). The \code{\link[=get_diver_data]{get_diver_data()}} function sets up an ODBC connection (using \code{\link[=db_connect]{db_connect()}}), which requires their quite limited \href{https://www.dimins.com/online-help/workbench_help/Content/ODBC/di-odbc.html}{DI-ODBC driver}.
}
\details{
These functions return a 'certedb tibble' from Diver or MOLIS, which prints information in the tibble header about the used source and current user.

Use \code{\link[=certedb_query]{certedb_query()}} to retrieve the original query that was used to download the data.

Using \code{certedb_query("your qry here")} is identical to using \code{certedb_getmmb(query = "your qry here")}.
}
\examples{
\dontrun{

# these two work identical:
get_diver_data(date_range = 2023, where = BepalingCode == "PXNCOV")
get_diver_data(2023, BepalingCode == "PXNCOV")

# use di$ to pull a list with column names while you type
get_diver_data(2023, di$BepalingCode == "PXNCOV")

# for the `where`, use `&`, `|`, or `c()`:
get_diver_data(last_month(),
               di$BepalingCode == "PXNCOV" & di$Zorglijn == "2e lijn")
get_diver_data(c(2020:2023),
               where = c(di$BepalingCode == "PXNCOV", di$Zorglijn == "2e lijn"))


# use \%like\%, \%unlike\%, \%like_case\% or \%unlike_case\% for regular expressions
get_diver_data(2023, where = di$MateriaalNaam \%like\% "Bloed")
get_diver_data(2023, where = di$MateriaalNaam \%unlike\% "Bloed")
get_diver_data(2023, where = di$MateriaalNaam \%like_case\% "bloed")
get_diver_data(2023, where = di$MateriaalNaam \%unlike_case\% "Bloed")

get_diver_data(2023,
               where = c(di$BepalingNaam \%like\% "Noro",
                         di$PatientLeeftijd >= 75))

# USING DIVER INTEGRATOR LANGUAGE --------------------------------------

# Use Diver Integrator functions within EVAL():              
get_diver_data(2023, where = EVAL('regexp(value("MateriaalCode"),"^B")'))

get_diver_data(
  2023,
  where = EVAL('rolling(12, value("OntvangstDatum"), date("2023/11/27"))')
)

# See the website for an overview of allowed functions:
# https://www.dimins.com/online-help/workbench_help/Content/ODBC/di-odbc-sql-reference.html
}


# USING DUCKDB DATABASE ------------------------------------------------

# create a local duckdb database and write a table
db <- db_connect(duckdb::duckdb(), "~/my_duck.db")
db |> db_write_table("iris", values = iris)
db |> db_close()

df <- get_duckdb_data(date_range = NULL,
                      where = Species == "setosa" & Sepal.Width > 3,
                      # not needed in production environment:
                      duckdb_path = "~/my_duck.db",
                      duckdb_table = "iris",
                      review_qry = FALSE,
                      info = TRUE)
df
certedb_query(df)

# remove the database
unlink("~/my_duck.db")
}
